feat(solver): Compute power density
This is rather ad-hoc, but I am not well versed enough in FEEC to figure
out what the proper dance here is. I need an inner product, but that
is created from a hodge star, except that lives on the dual mesh, so
it does not type match properly etc etc.