fix(cgal): Construct Polygon_2 properly
Shapely shapes contain the "first" coordinate at the start and at the
end of the coordinate array. However, CGAL Polygon_2 is implicitly
closed, so it makes no sense to include them twice.

This was causing is_simple_2 check to fail. Not yet relevant as of this
commit, but it is eventually going to be relevant when assertions are
enabled.